Fix columns without unit layers and height layout
