+++ frontend/scripts/react-components/tool/tool-modal/layer-modal/layer-modal.component.jsx
@@ -1,63 +1,13 @@
 import React, { useState, useMemo } from 'react';
 import PropTypes from 'prop-types';
 import Heading from 'react-components/shared/heading';
-import GridList from 'react-components/shared/grid-list/grid-list.component';
-import GridListItem from 'react-components/shared/grid-list-item/grid-list-item.component';
 import Text from 'react-components/shared/text';
 import Button from 'react-components/shared/button';
 import Tabs from 'react-components/shared/tabs/tabs.component';
 import { LAYER_TAB_NAMES } from 'constants';
 import 'react-components/tool/tool-modal/layer-modal/layer-modal.scss';
 import castArray from 'lodash/castArray';
-
-const renderList = (items, currentSelection, selectedTab, changeSelection) => {
-  const COLUMN_COUNT = 3;
-  const selectedItemIds = currentSelection[selectedTab];
-  const idAttribute = {
-    [LAYER_TAB_NAMES.unit]: 'uid',
-    [LAYER_TAB_NAMES.contextual]: 'id'
-  }[selectedTab];
-  const enableItem = item =>
-    changeSelection({
-      ...currentSelection,
-      [selectedTab]: [...(selectedItemIds || []), item[idAttribute]]
-    });
-  const disableItem = item =>
-    changeSelection({
-      ...currentSelection,
-      [selectedTab]: selectedItemIds.filter(id => id !== item[idAttribute])
-    });
-  const isActive = itemProps =>
-    selectedTab === LAYER_TAB_NAMES.unit
-      ? itemProps.item && !itemProps.item.isGroup && selectedItemIds?.includes(itemProps.item.uid)
-      : selectedItemIds?.includes(itemProps.item.id);
-  return (
-    <GridList
-      items={items}
-      height={items.length > COLUMN_COUNT ? (items.length / COLUMN_COUNT) * 60 + 20 : 100}
-      width={750}
-      rowHeight={50}
-      columnWidth={240}
-      columnCount={COLUMN_COUNT}
-      groupBy={selectedTab === LAYER_TAB_NAMES.unit ? 'group' : undefined}
-    >
-      {itemProps => (
-        <GridListItem
-          {...itemProps}
-          tooltip={itemProps.item?.description}
-          isActive={isActive(itemProps)}
-          enableItem={enableItem}
-          disableItem={disableItem}
-          isDisabled={
-            selectedTab === LAYER_TAB_NAMES.unit &&
-            !isActive(itemProps) &&
-            selectedItemIds?.length > 1
-          }
-        />
-      )}
-    </GridList>
-  );
-};
+import LayersList from 'react-components/tool/tool-modal/layer-modal/layers-list';
 
 export default function LayerModal({
   layers,
@@ -102,8 +52,8 @@ export default function LayerModal({
   }[selectedTab];
   return (
     <div className="c-layer-modal">
-      <div className="layer-modal-content">
-        <div className="row columns">
+      <div className="row columns">
+        <div className="layer-modal-content">
           <Heading size="md" className="modal-title">
             Edit map layers
           </Heading>
@@ -111,8 +61,14 @@ export default function LayerModal({
           <Text color="grey-faded" size="md" className="info-message">
             {infoMessage}
           </Text>
-          {layers[selectedTab] &&
-            renderList(layers[selectedTab], currentSelection, selectedTab, changeSelection)}
+          {layers[selectedTab] && (
+            <LayersList
+              items={layers[selectedTab]}
+              currentSelection={currentSelection}
+              selectedTab={selectedTab}
+              changeSelection={changeSelection}
+            />
+          )}
         </div>
       </div>
       <div className="modal-footer">
