+++ frontend/scripts/react-components/tool/tool-modal/layer-modal/layers-list/layers-list.js
@@ -1,0 +1,3 @@
+export {
+  default
+} from 'react-components/tool/tool-modal/layer-modal/layers-list/layers-list.component';
